pyproject.toml
[tool.poetry]
authors = ["govindsalinas"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
pangea-sdk = "^5.3.0"
kong-pdk = "^0.36"

